forgot pwd page, insertion of question+ image to db and file

+++ imageupload.php
@@ -1,31 +1,39 @@
 <?php
+   session_start();
+   if(isset($_SESSION['key']) && $_SESSION['key'] =="9FOj92VfeSbTQ54M" ){
+      $subid= isset($_POST['subid'])?$_POST['subid']:'';
+      $name= isset($_POST['name'])?$_POST['name']:'';
+      
+      /* Getting file name */
+      $filename = $_FILES['file']['name'];
+      //$filename = $name;
 
-/* Getting file name */
-$filename = $_FILES['file']['name'];
+     
+      $uploadOk = 1;
+      $imageFileType = pathinfo($filename,PATHINFO_EXTENSION);
+       /* Location */
+       $location = "questionImages/".$subid."/".$name;
+       $location.='.jpg';
 
-/* Location */
-$location = "questionImages/".$filename;
-$uploadOk = 1;
-$imageFileType = pathinfo($location,PATHINFO_EXTENSION);
-/* Valid Extensions */
-$valid_extensions = array("jpg","jpeg","png");
-/* Check file extension */
-if( !in_array(strtolower($imageFileType),$valid_extensions) ) {
-   $uploadOk = 0;
-}
+      /* Valid Extensions */
+      $valid_extensions = array("jpg","jpeg","png");
+      /* Check file extension */
+      if( !in_array(strtolower($imageFileType),$valid_extensions) ) {
+         $uploadOk = 0;
+      }
 
-if($uploadOk == 0){
-   echo "valid format";
-}else{
-   /* Upload file */
-
-   echo "file type ".$location;
-   if(file_exists($filename))
-    unlink($filename);
-    echo "fiel".$_FILES['file']['tmp_name'];
-   if(move_uploaded_file($_FILES['file']['tmp_name'],$location)){
-      echo $location;
-   }else{
-      echo 0;
+      if($uploadOk == 0){
+         echo 0;
+      }else{
+         /* Upload file */
+         if(file_exists($filename))
+            unlink($filename);
+         //echo "file".$_FILES['file']['tmp_name'];
+         if(move_uploaded_file($_FILES['file']['tmp_name'],$location)){
+            echo 1;
+         }else{
+            echo 0;
+         }
+      }
    }
-}
+?>
